// libs/core/src/lib/config/env/names.enum.ts
export enum ENV_NAME {
  // System
  STORAGE_DEST = 'STORAGE_DEST',
  MONGO_URI = 'MONGO_URI',
  PORT_API = 'PORT_API',
  HOST_API = 'HOST_API',

  // Swaggger configuration
  SWAGGER_ENABLED = 'SWAGGER_ENABLED',
  SWAGGER_TITLE = 'SWAGGER_TITLE',
  SWAGGER_DESC = 'SWAGGER_DESC',
  SWAGGER_PATH = 'SWAGGER_PATH',
  SWAGGER_USER = 'SWAGGER_USER',
  SWAGGER_PASSWORD = 'SWAGGER_PASSWORD',

  // Security
  CORS_ENABLED = 'CORS_ENABLED',
  HELMET_ENABLED = 'HELMET_ENABLED',
  JWT_EXPIRES_IN = 'JWT_EXPIRES_IN',
  JWT_REFRESH_IN = 'JWT_REFRESH_IN',
  BCRYPT_SALT_ROUND = 'BCRYPT_SALT_ROUND',
  PASSWORD_RESET_TOKEN_EXPIRES_IN = 'PASSWORD_RESET_TOKEN_EXPIRES_IN',
}